extracted pcodeOpSimple into seperate file
